Raspberry Pi/requirements.txt
attrs==25.3.0
av==14.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
libarchive-c==5.2
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pillow==11.2.1
pyserial==3.5
pytesseract==0.3.13
python-prctl==1.8.1
referencing==0.36.2
rpds-py==0.24.0
simplejpeg==1.8.2
tqdm==4.67.1
typing_extensions==4.13.2
v4l2-python3==0.3.5
Werkzeug==3.1.3
